requirements.txt
Flask
Flask-SQLAlchemy
Flask-WTF
Flask-Login
psycopg2-binary
python-dotenv
email_validator
fab-coreui-theme
flask-socketio[asyncio]
gunicorn
eventlet